Demos/requirements.txt
boto3
aws-cdk-lib
constructs>=10.0.0,<11.0.0
requests
cmake
pandas
awswrangler
flask
logging